renamed setupwindow to scriptui
The package contains more than only the setup window.
