+++ phoenicis-cli/src/main/java/org/phoenicis/cli/scriptui/CliMessageSender.java
@@ -16,7 +16,7 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package org.phoenicis.cli.setupwindow;
+package org.phoenicis.cli.scriptui;
 
 import org.phoenicis.scripts.ui.UiMessageSender;
