+++ phoenicis-cli/src/main/java/org/phoenicis/cli/scriptui/SetupUiCliImplementation.java
@@ -16,7 +16,7 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package org.phoenicis.cli.setupwindow;
+package org.phoenicis.cli.scriptui;
 
 import org.jsoup.Jsoup;
 import org.phoenicis.scripts.ui.MenuItem;
