+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/scriptui/ProgressUiJavaFXImplementation.java
@@ -16,7 +16,7 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package org.phoenicis.javafx.views.setupwindow;
+package org.phoenicis.javafx.views.scriptui;
 
 import javafx.application.Platform;
 import javafx.scene.control.ProgressBar;
