+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/scriptui/StepRepresentationSpin.java
@@ -16,7 +16,7 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package org.phoenicis.javafx.views.setupwindow;
+package org.phoenicis.javafx.views.scriptui;
 
 import org.phoenicis.scripts.ui.Message;
 import javafx.scene.control.ProgressIndicator;
