+++ phoenicis-javafx/src/main/resources/org/phoenicis/javafx/views/scriptui/defaultTopImage.png
