+++ phoenicis-javafx/src/main/resources/org/phoenicis/javafx/views/scriptui/spin/3.png
