+++ phoenicis-javafx/src/main/resources/org/phoenicis/javafx/views/scriptui/spin/4.png
