+++ phoenicis-javafx/src/main/resources/org/phoenicis/javafx/views/scriptui/spin/7.png
